+++ core/index.js
@@ -25,9 +25,7 @@ export * from './HoistComponent';
 export * from './HoistModel';
 export * from './HoistService';
 
-export * from './task/TaskObserver';
-export * from './task/PromiseTaskObserver';
-export * from './task/CompoundTaskObserver';
+export * from './TaskObserver';
 
 export * from './refresh/LoadSupport';
 export * from './refresh/RefreshContextModel';