+++ data/impl/RecordFieldValidator.js
@@ -5,7 +5,7 @@
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
 
-import {HoistBase, managed, PromiseTaskObserver} from '@xh/hoist/core';
+import {HoistBase, managed, TaskObserver} from '@xh/hoist/core';
 import {ValidationState} from '@xh/hoist/data';
 import {computed, makeObservable, observable, runInAction} from '@xh/hoist/mobx';
 import {compact, flatten, isEmpty, isNil} from 'lodash';
@@ -72,7 +72,7 @@ export class RecordFieldValidator extends HoistBase {
     // completed will contain null
     @observable _errors;
 
-    @managed _validationTask = new PromiseTaskObserver();
+    @managed _validationTask = new TaskObserver.trackLast();
     _validationRunId = 0;
 
     /**